moved lensMaker to its own file
