+++ app/src/index.ts
@@ -1,5 +1,4 @@
 import { Lens } from './Lens'
+export { createLensMaker } from './createLensMaker'
 
 export type LensType = typeof Lens
-
-export default Lens
