+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "target": "es6",
-        "module": "commonjs" /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */,
+        "module": "commonjs",
         "lib": ["es5", "es6"],
         "declaration": true,
         "declarationMap": true,
@@ -19,7 +19,6 @@
         "noImplicitReturns": true,
         "moduleResolution": "node",
         "typeRoots": ["node_modules/@types", "types"],
-        // "esModuleInterop": true, /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
         "forceConsistentCasingInFileNames": true,
         "preserveConstEnums": true,
         "resolveJsonModule": true
